requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.11
pandas==2.2.3
openpyxl==3.1.5
werkzeug==3.1.3
pytz==2024.2
python-dotenv==1.0.1
WTForms==3.2.1
email-validator==2.2.0
WeasyPrint==63.1
matplotlib>=3.9.4
qrcode==8.0
Pillow>=11.0.0
